fix enum namespace Figure and CoordinateError
