+++ CoordinateCalculator/CoordinateCalculator/MyLine.swift
@@ -9,12 +9,12 @@
 import Foundation
 
 struct MyLine: MyShape {
-    var type: Figure
+    var type: OutputView.Figure
     private var pointA = MyPoint(x: 0, y: 0)
     private var pointB = MyPoint(x: 0, y: 0)
 
     init(pointA: MyPoint, pointB: MyPoint) {
-        self.type = Figure.line
+        self.type = OutputView.Figure.line
         self.pointA = pointA
         self.pointB = pointB
     }
