+++ CoordinateCalculator/CoordinateCalculator/MyPoint.swift
@@ -9,16 +9,18 @@
 import Foundation
 
 struct MyPoint: MyShape {
-    var type: Figure
+    
+    var type: OutputView.Figure
     private(set) var x = 0
     private(set) var y = 0
     init(x: Int, y: Int) {
-        self.type = Figure.point
+        self.type = OutputView.Figure.point
         self.x = x
         self.y = y
     }
     
-    func calculatorOfPosition() -> String{
+    func calculateOfPosition() -> String {
         return "\(ANSICode.text.redBright)\(ANSICode.cursor.move(row: 25 - y, col: (x * 2) + 3))\(ANSICode.text.dot)"
     }
+
 }
